move AlignThreadsIcon to CoreFixes
